feat: #18 Answerコンポーネント上の正解判定ロジックをAppに移した
