+++ src/components/answer.tsx
@@ -1,15 +1,6 @@
-import React, { useState, useEffect } from "react";
-import axios from "axios";
-
 type Props = {
   answerList: string[][];
   matchList: string[][];
-  setMatchList: React.Dispatch<React.SetStateAction<string[][]>>;
-  judge: boolean;
-  setJudge: React.Dispatch<React.SetStateAction<boolean>>;
-  correctAnswer: string;
-  gameStatus: string;
-  setGameStatus: React.Dispatch<React.SetStateAction<string>>;
 };
 
 export const Answer = (props: Props) => {
@@ -62,123 +53,6 @@ export const Answer = (props: Props) => {
     "Yellow": yellowTdStyle,
     "Green": greenTdStyle,
   };
-
-  // ラウンド
-  const [round, setRound] = useState<number>(0);
-
-  // 単語の妥当性判定
-  const wordValidityJudgement = async () => {
-    const { data } = await axios.post('https://yan5p8s0dg.execute-api.ap-southeast-2.amazonaws.com/WORDLE', 
-      {"word": props.answerList[round - 1].join("")},);
-    console.log(data);
-    if (data.isValid === undefined) {
-      return false;
-    }
-
-    return data.isValid;
-  }
-
-  // 単語一致判定
-  const wordMatchJudgement = (): string[][] => {
-    // 一度ディープコピーする
-    const tmpMatchList = Array.from(props.matchList);
-
-    // 1文字ずつ判定
-    for (let i = 0; i < 5; i++) {
-      // 文字が一致
-      if (props.correctAnswer.indexOf(props.answerList[round - 1][i]) !== -1) {
-        // 位置も一致(Green)
-        if (props.answerList[round - 1][i] === props.correctAnswer[i]) {
-          tmpMatchList[round - 1][i] = "Green";
-        }
-
-        // 文字だけ一致(Yellow)
-        else {
-          tmpMatchList[round - 1][i] = "Yellow";
-        }
-      }
-
-      // 文字も位置も一致していない(Black)
-      else {
-        tmpMatchList[round - 1][i] = "Black";
-      }
-    }
-    return tmpMatchList;
-  };
-
-  // クリア判定
-  const clearJudgement = () => {
-    // 正解が空文字だった場合、alertを出す
-    if (props.correctAnswer === ""){
-        alert("Server Error: Please reload the page.");
-        return;
-    }
-
-    // ワードを抽出
-    const wordList = [];
-    for (let j = 0; j < 5; j++) {
-      wordList.push(props.answerList[round - 1][j]);
-    }
-    const submitWord = wordList.join("");
-
-    if (submitWord == props.correctAnswer) {
-      alert("clear!!");
-      return "success";
-    } else if (round == 6) {
-      alert(props.correctAnswer);
-      return "fail";
-    }
-
-    return "playing";
-  };
-
-  // Appコンポーネントのjudgeが変化した時に呼ばれる
-  useEffect(() => {
-
-    const checkProcess = async () => {
-
-      // Enterを押したら
-      if (props.judge === true) {
-
-        // 単語の妥当性判定
-        const isValid = await wordValidityJudgement();
-        if (!isValid)
-        {
-          alert("データセットに存在しない単語です");
-          return;
-        }
-
-        // 一度フラグをおろす
-        props.setJudge(false);
-      }
-
-      // フラグをおろしてからここへ
-      else {
-        // コンポーネント初期化時にここを通る
-        if (round == 0) {
-          setRound(round + 1); // ラウンドを1に
-          return;
-        }
-
-        // ゲーム継続中なら
-        if (props.gameStatus == "playing") {
-          // 単語一致判定
-          // const tmpMatchStyleList = wordMatchJudgement();
-          const tmpMatchList = wordMatchJudgement();
-          // クリア判定
-          clearJudgement();
-          // スタイル更新
-          // setMatchStyleList(tmpMatchStyleList);
-          props.setMatchList(tmpMatchList);
-          // ラウンド更新
-          setRound(round + 1);
-        }
-      }
-    }
-
-    checkProcess();
-
-  }, [props.judge]);
 
   return (
     // mapにより回答table作成
